Organized classes into correct folders and namespaces.
