+++ Sudokungfu/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Sudokungfu"
-             StartupUri="SudokuWindow.xaml">
+             StartupUri="View/Sudokungfu.xaml">
     <Application.Resources>
          
     </Application.Resources>
