+++ Sudokungfu/SudokuSolver/Techniques/PossibleSpotOverlapTechnique.cs
@@ -55,15 +55,14 @@ namespace Sudokungfu.SudokuSolver.Techniques
         /// <param name="value">Value that must go in the overlapping cells.</param>
         /// <param name="indexes">Indexes of the overlapping cells.</param>
         /// <param name="setindexes">Indexes of the cells in the sets that overlap.</param>
-        /// <param name="affectedIndexes">Indexes of cells that had values eliminanated by this technique.</param>
+        /// <param name="affectedIndexes">Indexes of cells that had values eliminated by this technique.</param>
         public static PossibleSpotOverlapTechnique CreatePossibleSpotOverlapTechnique(int value, IEnumerable<int> indexes, IEnumerable<int> setindexes, IEnumerable<int> affectedIndexes)
         {
             return new PossibleSpotOverlapTechnique()
             {
                 Complexity = DEFAULT_COMPLEXITY,
                 IndexValueMap = setindexes.ToDictionary(i => i, i => indexes.Contains(i) ? value.ToEnumerable() : Enumerable.Empty<int>()),
-                AffectedIndexes = affectedIndexes,
-                UsesFoundValues = false
+                AffectedIndexes = affectedIndexes
             };
         }
     }
