+++ Sudokungfu/SudokuSolver/Techniques/ThreeSpotClosureTechnique.cs
@@ -63,7 +63,7 @@ namespace Sudokungfu.SudokuSolver.Techniques
         /// <param name="valueB">Value in the closure and cells which the value can go in.</param>
         /// <param name="valueC">Value in the closure and cells which the value can go in.</param>
         /// <param name="setIndexes">Indexes of the cells in the set with a closure.</param>
-        /// <param name="affectedIndexes">Indexes of cells that had values eliminanated by this technique.</param>
+        /// <param name="affectedIndexes">Indexes of cells that had values eliminated by this technique.</param>
         public static ThreeSpotClosureTechnique CreateThreeSpotClosureTechnique(KeyValuePair<int, IEnumerable<int>> valueA, KeyValuePair<int, IEnumerable<int>> valueB, KeyValuePair<int, IEnumerable<int>> valueC, IEnumerable<int> setIndexes, IEnumerable<int> affectedIndexes)
         {
             var values = new KeyValuePair<int, IEnumerable<int>>[] { valueA, valueB, valueC };
@@ -71,8 +71,7 @@ namespace Sudokungfu.SudokuSolver.Techniques
             {
                 Complexity = DEFAULT_COMPLEXITY,
                 IndexValueMap = setIndexes.ToDictionary(i => i, i => values.Where(v => v.Value.Contains(i)).Select(v => v.Key)),
-                AffectedIndexes = affectedIndexes,
-                UsesFoundValues = false
+                AffectedIndexes = affectedIndexes
             };
         }
     }
