+++ Sudokungfu/View/Sudokungfu.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Sudokungfu.SudokuWindow"
+﻿<Window x:Class="Sudokungfu.View.Sudokungfu"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Sudokungfu"
+        xmlns:local="clr-namespace:Sudokungfu.View"
         xmlns:prop="clr-namespace:Sudokungfu.Properties"
         mc:Ignorable="d"
         Title="{x:Static prop:Resources.SudokuWindowTitle}"
